requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
groq==1.1.2
pymongo==4.16.0
dnspython
python-dotenv==1.2.2
python-multipart==0.0.22
twilio==9.10.4
ultralytics==8.4.32
opencv-python-headless==4.13.0.92
xhtml2pdf==0.2.17
numpy==2.4.4
pillow==12.1.1
